+++ server/api.js
@@ -23,111 +23,7 @@ const auth = require("./auth");
 const router = express.Router();
 
 const socketManager = require("./server-socket");
-
-// anyscale setup
-const ANYSCALE_API_KEY = "";
-// some information about this model: https://ai.meta.com/llama/
-const MODEL = "meta-llama/Llama-2-13b-chat-hf";
-// another common choice is text-embedding-ada-002.
-// we use gte-large because this is the only embedding model anyscale has access to
-const EMBEDDING_MODEL = "thenlper/gte-large";
-const { OpenAI } = require("openai");
-const anyscale = new OpenAI({
-  baseURL: "https://api.endpoints.anyscale.com/v1",
-  apiKey: ANYSCALE_API_KEY,
-});
-
-// embedding helper function
-const generateEmbedding = async (document) => {
-  const embedding = await anyscale.embeddings.create({
-    model: EMBEDDING_MODEL,
-    // model: "text-embedding-ada-002", (anyscale doesn't have access to this model)
-    input: document,
-  });
-  return embedding.data[0].embedding;
-};
-
-// chat completion helper function
-const chatCompletion = async (query, context) => {
-  const prompt = {
-    model: MODEL,
-    messages: [
-      {
-        role: "system",
-        content:
-          "Your role is to answer questions for a user. You are given the following context to help you answer questions: \n" +
-          `${context}. \n` +
-          "Please do not mention that you were given any context in your response.",
-      },
-      { role: "user", content: `${query}` },
-    ],
-    // temperature controls the variance in the llms responses
-    // higher temperature = more variance
-    temperature: 0.7,
-  };
-  const completion = await anyscale.chat.completions.create(prompt);
-  return completion.choices[0].message.content;
-};
-
-// initialize vector database
-const COLLECTION_NAME = "catbook-collection";
-const { ChromaClient } = require("chromadb");
-const client = new ChromaClient();
-
-let collection;
-
-// sync main and vector dbs
-const syncDBs = async () => {
-  // retrieve all documents
-  const allDocs = await collection.get();
-  // delete all documents
-  await collection.delete({
-    ids: allDocs.ids,
-  });
-  // retrieve corpus from main db
-  const allMongoDocs = await Document.find({});
-  const allMongoDocIds = allMongoDocs.map((mongoDoc) => mongoDoc._id.toString());
-  const allMongoDocContent = allMongoDocs.map((mongoDoc) => mongoDoc.content);
-  let allMongoDocEmbeddings = allMongoDocs.map((mongoDoc) => generateEmbedding(mongoDoc.content));
-  allMongoDocEmbeddings = await Promise.all(allMongoDocEmbeddings); // ensure embeddings finish generating
-  // add corpus to vector db
-  await collection.add({
-    ids: allMongoDocIds,
-    embeddings: allMongoDocEmbeddings,
-    documents: allMongoDocContent,
-  });
-  console.log("number of documents", await collection.count());
-  console.log("finished initializing chroma collection");
-};
-
-const initCollection = async () => {
-  collection = await client.getOrCreateCollection({
-    name: COLLECTION_NAME,
-  });
-  // initialize collection embeddings with corpus
-  // in production, this function should not run that often, so it is OK to resync the two dbs here
-  await syncDBs();
-};
-
-initCollection();
-
-// retrieving context helper function
-const NUM_DOCUMENTS = 2;
-const retrieveContext = async (query, k) => {
-  const queryEmbedding = await generateEmbedding(query);
-  const results = await collection.query({
-    queryEmbeddings: [queryEmbedding],
-    nResults: k,
-  });
-  return results.documents;
-};
-
-// RAG
-const retrievalAugmentedGeneration = async (query) => {
-  const context = await retrieveContext(query, NUM_DOCUMENTS);
-  const llmResponse = await chatCompletion(query, context);
-  return llmResponse;
-};
+const ragManager = require("./rag");
 
 router.get("/stories", (req, res) => {
   // empty selector means get all documents
@@ -159,70 +55,6 @@ router.post("/comment", auth.ensureLoggedIn, (req, res) => {
   });
 
   newComment.save().then((comment) => res.send(comment));
-});
-
-router.post("/document", (req, res) => {
-  const newDocument = new Document({
-    content: req.body.content,
-  });
-
-  const addDocument = async (document) => {
-    try {
-      await document.save();
-      const embedding = await generateEmbedding(document.content);
-      await collection.add({
-        ids: [document._id.toString()],
-        embeddings: [embedding],
-        documents: [document.content],
-      });
-      res.send(document);
-    } catch (error) {
-      console.log("error:", error);
-      res.status(500);
-      res.send({});
-    }
-  };
-
-  addDocument(newDocument);
-});
-
-router.get("/document", (req, res) => {
-  retrieveContext("hello", NUM_DOCUMENTS).then(() => {
-    Document.find({}).then((documents) => res.send(documents));
-  });
-});
-
-router.post("/updateDocument", (req, res) => {
-  const updateDocument = async (id) => {
-    const document = await Document.findById(id);
-    document.content = req.body.content;
-    await document.save();
-    await collection.modify({
-      ids: [document._id],
-      documents: [document.content],
-    });
-    res.send({});
-  };
-  updateDocument(req.body._id);
-});
-
-router.post("/deleteDocument", (req, res) => {
-  const deleteDocument = async (id) => {
-    const document = await Document.findById(id);
-    if (!document) res.send({});
-    try {
-      await collection.delete({
-        ids: [document._id],
-      });
-      await document.remove();
-      res.send({});
-    } catch {
-      // if deleting from the vector db failed (e.g., it doesn't exist)
-      await document.remove();
-      res.send({});
-    }
-  };
-  deleteDocument(req.body._id);
 });
 
 router.post("/login", auth.login);
@@ -309,10 +141,62 @@ router.post("/despawn", (req, res) => {
   res.send({});
 });
 
+router.post("/document", (req, res) => {
+  const newDocument = new Document({
+    content: req.body.content,
+  });
+
+  const addDocument = async (document) => {
+    try {
+      await document.save();
+      await ragManager.addDocument(document);
+      res.send(document);
+    } catch (error) {
+      console.log("error:", error);
+      res.status(500);
+      res.send({});
+    }
+  };
+
+  addDocument(newDocument);
+});
+
+router.get("/document", (req, res) => {
+  Document.find({}).then((documents) => res.send(documents));
+});
+
+router.post("/updateDocument", (req, res) => {
+  const updateDocument = async (id) => {
+    const document = await Document.findById(id);
+    document.content = req.body.content;
+    await document.save();
+    await ragManager.updateDocument(document);
+    res.send({});
+  };
+  updateDocument(req.body._id);
+});
+
+router.post("/deleteDocument", (req, res) => {
+  const deleteDocument = async (id) => {
+    const document = await Document.findById(id);
+    if (!document) res.send({});
+    try {
+      await ragManager.deleteDocument(id);
+      await document.remove();
+      res.send({});
+    } catch {
+      // if deleting from the vector db failed (e.g., it doesn't exist)
+      await document.remove();
+      res.send({});
+    }
+  };
+  deleteDocument(req.body._id);
+});
+
 router.post("/query", (req, res) => {
   const makeQuery = async () => {
     try {
-      const queryresponse = await retrievalAugmentedGeneration(req.body.query);
+      const queryresponse = await ragManager.retrievalAugmentedGeneration(req.body.query);
       res.send({ queryresponse });
     } catch (error) {
       console.log("error:", error);
